PT-4733 - Use article mode from position for order detail
Previously, all order details were set to "modus 4 = payment surcharge
/ discount".

This had the effect that no positions from orders created with this
plugin show up on index_ls.tpl (Delivery note), because this template
only outputs order details with modes 0 or 1 (default or premium
articles).

Now, the actual position's mode is used to set the mode of the order
detail.