PT-4736 - Fix automatic shipping address selection
When a user deselects the "Use billing address [for shipping]" checkbox,
the first entry of the shipping address combobox is automatically
selected. However, this had no further effect.

Now, automatically selected entries for a distinct shipping address
are actually persisted. Previously, neither would the view update
correspondingly, nor would the client store such a shippingAddressId in
the model (and hence also would not use it for creating the order).